refactor(cli): extract shared helpers to cli/shared/ to break bidirectional cycle
Move types.py, helpers.py, helper_support.py into cli/shared/ so that
cli/commands/ → cli/shared/ → (no back-edge to commands). This resolves
the cli/ ↔ cli/commands/ import cycle.

Ref #431
