+++ polylogue/cli/bulk_export.py
@@ -10,7 +10,7 @@ from polylogue.api.sync.bridge import run_coroutine_sync
 from polylogue.cli.query import project_query_results
 from polylogue.cli.query_contracts import QueryExecutionPlan
 from polylogue.cli.root_request import RootModeRequest
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.types import AppEnv
 from polylogue.rendering.formatting import format_conversation
 
 _PER_LINE_FORMATS = frozenset({"jsonl"})
