+++ polylogue/cli/check_maintenance.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from polylogue.cli.check_models import VacuumResult
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.types import AppEnv
 from polylogue.maintenance.targets import MaintenanceTargetMode, build_maintenance_target_catalog
 from polylogue.readiness import ReadinessReport
 from polylogue.storage.repair import RepairResult, preview_counts_from_archive_debt
