+++ polylogue/cli/check_rendering_plain.py
@@ -7,7 +7,7 @@ import click
 from polylogue.cli.check_models import CheckCommandResult
 from polylogue.cli.check_support import format_count_mapping, run_vacuum
 from polylogue.cli.check_workflow import CheckCommandOptions
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.types import AppEnv
 from polylogue.readiness import VerifyStatus
 
 # ---------------------------------------------------------------------------
