+++ polylogue/cli/check_support.py
@@ -6,8 +6,8 @@ import sys
 import time
 
 from polylogue.cli.check_models import VacuumResult
-from polylogue.cli.helpers import fail
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.helpers import fail
+from polylogue.cli.shared.types import AppEnv
 from polylogue.protocols import ProgressCallback
