+++ polylogue/cli/click_app.py
@@ -22,7 +22,7 @@ from polylogue.cli.machine_main import extract_option as _extract_option
 from polylogue.cli.machine_main import run_machine_entry
 from polylogue.cli.query import QueryFirstGroupBase, handle_query_mode
 from polylogue.cli.query_verbs import QUERY_VERBS
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.types import AppEnv
 from polylogue.logging import configure_logging
 from polylogue.ui import create_ui
 from polylogue.version import POLYLOGUE_VERSION
@@ -42,7 +42,7 @@ def _handle_query_mode(ctx: click.Context) -> None:
 
 def _show_stats(env: AppEnv, *, verbose: bool = False) -> None:
     """Show archive statistics."""
-    from polylogue.cli.helpers import print_summary
+    from polylogue.cli.shared.helpers import print_summary
 
     print_summary(env, verbose=verbose)
