+++ polylogue/cli/commands/export.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 import click
 
 from polylogue.api.sync.bridge import run_coroutine_sync
-from polylogue.cli.helper_support import fail
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.helper_support import fail
+from polylogue.cli.shared.types import AppEnv
 from polylogue.lib.message.roles import normalize_message_roles
 from polylogue.rendering.formatting import CONVERSATION_OUTPUT_FORMATS, format_conversation
