+++ polylogue/cli/commands/qa.py
@@ -6,13 +6,13 @@ from pathlib import Path
 
 import click
 
-from polylogue.cli.helpers import complete_configured_source_names, load_effective_config, resolve_sources
 from polylogue.cli.qa_finalization import finalize_qa_run
 from polylogue.cli.qa_requests import (
     build_qa_invocation_plan,
 )
 from polylogue.cli.qa_snapshot import execute_snapshot_plan
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.helpers import complete_configured_source_names, load_effective_config, resolve_sources
+from polylogue.cli.shared.types import AppEnv
 from polylogue.showcase.qa_runner_request import QAStage
 
 _PRODUCT_STAGE_CHOICES = click.Choice([QAStage.AUDIT.value])
