+++ polylogue/cli/qa_finalization.py
@@ -13,7 +13,7 @@ from polylogue.cli.qa_requests import QACaptureMode, QAFinalizationPlan
 from polylogue.cli.qa_snapshot import execute_snapshot_plan
 
 if TYPE_CHECKING:
-    from polylogue.cli.types import AppEnv
+    from polylogue.cli.shared.types import AppEnv
     from polylogue.showcase.qa_runner_models import QAResult
