+++ polylogue/cli/qa_snapshot.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 from polylogue.cli.machine_errors import emit_success
 from polylogue.cli.qa_requests import QASnapshotPlan
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.types import AppEnv
 from polylogue.paths.sanitize import safe_path_component
