+++ polylogue/cli/query_actions.py
@@ -10,7 +10,7 @@ import click
 from polylogue.cli.query_contracts import QueryMutationSpec, result_date, result_id, result_provider, result_title
 
 if TYPE_CHECKING:
-    from polylogue.cli.types import AppEnv
+    from polylogue.cli.shared.types import AppEnv
     from polylogue.lib.filter.filters import ConversationFilter
     from polylogue.lib.models import Conversation, ConversationSummary
     from polylogue.lib.query.spec import ConversationQuerySpec
