+++ polylogue/cli/query_feedback.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 from polylogue.cli.machine_errors import error_no_results
 
 if TYPE_CHECKING:
-    from polylogue.cli.types import AppEnv
+    from polylogue.cli.shared.types import AppEnv
     from polylogue.lib.query.miss_diagnostics import QueryMissDiagnostics
     from polylogue.lib.query.spec import ConversationQuerySpec
