+++ polylogue/cli/query_semantic.py
@@ -10,7 +10,7 @@ from polylogue.cli.query_feedback import emit_no_results
 from polylogue.cli.query_stats import emit_structured_stats
 
 if TYPE_CHECKING:
-    from polylogue.cli.types import AppEnv
+    from polylogue.cli.shared.types import AppEnv
     from polylogue.lib.action_event.action_events import ActionEvent
     from polylogue.lib.models import Conversation, ConversationSummary
     from polylogue.lib.query.spec import ConversationQuerySpec
