+++ polylogue/cli/shared/helpers.py
@@ -13,8 +13,8 @@ from polylogue.cli.helper_source_selection import (
 )
 from polylogue.cli.helper_source_state import load_last_source, save_last_source, source_state_path
 from polylogue.cli.helper_summary import print_summary_impl
-from polylogue.cli.helper_support import fail, load_effective_config
-from polylogue.cli.types import AppEnv
+from polylogue.cli.shared.helper_support import fail, load_effective_config
+from polylogue.cli.shared.types import AppEnv
 from polylogue.operations import get_provider_counts, list_provider_analytics_products
 from polylogue.pipeline.runner import latest_run
 from polylogue.readiness import get_readiness, quick_readiness_summary
